requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.3
attrs==22.2.0
autopep8==2.0.2
backcall==0.2.0
beautifulsoup4==4.12.3
black==24.4.2
bleach==6.0.0
certifi==2024.6.2
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
comm==0.1.3
dataclasses-json==0.6.7
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
executing==1.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
fastjsonschema==2.16.3
flake8==6.0.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.4
ipykernel==6.22.0
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.5
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.6
langchain==0.2.3
langchain-core==0.2.5
langchain-openai==0.1.8
langchain-text-splitters==0.2.1
langserve==0.2.1
langsmith==0.1.75
llama-index==0.10.47
llama-index-agent-openai==0.2.7
llama-index-cli==0.1.12
llama-index-core==0.10.47
llama-index-embeddings-openai==0.1.10
llama-index-experimental==0.1.3
llama-index-indices-managed-llama-cloud==0.1.6
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.22
llama-index-multi-modal-llms-openai==0.1.6
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-file==0.1.25
llama-index-readers-llama-parse==0.1.4
llama-parse==0.4.4
llamaindex-py-client==0.1.19
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.21.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==2.0.5
multidict==6.0.5
mypy-extensions==1.0.0
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.26.4
openai==1.32.0
orjson==3.10.3
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.2.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.13
pydantic_core==2.18.4
pyflakes==3.0.1
Pygments==2.14.0
pypdf==4.2.0
pyperclip==1.8.2
pyproject-toml==0.0.10
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.0.2
qtconsole==5.4.1
QtPy==2.3.0
regex==2024.5.15
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
Send2Trash==1.8.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.30
sse-starlette==1.8.2
stack-data==0.6.2
starlette==0.37.2
striprtf==0.0.26
tenacity==8.3.0
terminado==0.17.1
tiktoken==0.7.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
tqdm==4.66.4
traitlets==5.9.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.1
tzdata==2024.1
ujson==5.10.0
uri-template==1.2.0
urllib3==2.2.1
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==12.0
widgetsnbextension==4.0.6
wrapt==1.16.0
yarl==1.9.4
